BA-685: Split backend code to own folder
