+++ next/backend/utils/cookies.ts
@@ -2,11 +2,10 @@
 import { captureException } from '@sentry/nextjs'
 import mapValues from 'lodash/mapValues'
 import pick from 'lodash/pick'
+import { isBrowser } from '@utils/utils'
 import { useCallback, useEffect, useState } from 'react'
 // todo swap for js-cookie
 import { Cookies } from 'react-cookie-consent'
-
-import { isBrowser } from './utils'
 
 const availableConsents = ['statistics']
 const pickConsents = (consents: any) => mapValues(pick(consents, availableConsents), Boolean)
