+++ next/components/molecules/OrganizationalStructure/OrganizationalStructureAccordion.tsx
@@ -2,7 +2,7 @@
 import { roleOrderingScore } from '@utils/organisationalStructure'
 import cx from 'classnames'
 import { useToggle } from 'rooks'
-import { GetGroupMembersRecursiveResult } from 'services/ms-graph'
+import { GetGroupMembersRecursiveResult } from 'backend/services/ms-graph'
 
 import ChevronDown from '../../../assets/images/chevron-down-thin.svg'
 import ChevronDownSmall from '../../../assets/images/chevron-down-thin-small.svg'
