+++ next/components/molecules/sections/documentList/index.tsx
@@ -5,12 +5,12 @@ import DocumentDevider from '@assets/images/documentDevider.svg'
 import DocumentDeviderSmall from '@assets/images/documentDevider-small.svg'
 import { BasicSearch, DocumentListItem, Modal, NoResultsFound, Pagination } from '@bratislava/ui-bratislava'
 import DocumentListCategorysMap from '@utils/documentListCategory'
-import { MEILI_PAGE_SIZE, searchVZN } from '@utils/meili'
 import { fileCountVzns } from '@utils/utils'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
 import useSWR from 'swr'
 
+import { MEILI_PAGE_SIZE, searchVZN } from '../../../../backend/utils/meili'
 import { DocumentListModalBody } from './modalBody'
 
 export const DocumentList = () => {
