+++ next/pages/api/ms-graph/users/search/[query].tsx
@@ -1,9 +1,9 @@
-//  https://graph.microsoft.com/v1.0/users?$search="displayName:Kmet"
+/// /https://graph.microsoft.com/v1.0/users?$search="displayName:Kmet"
 
 import { withSentry } from '@sentry/nextjs'
 import { forceString } from '@utils/utils'
+import { getToken, getUsersByDisplayName } from 'backend/services/ms-graph'
 import type { NextApiRequest, NextApiResponse } from 'next'
-import { getToken, getUsersByDisplayName } from 'services/ms-graph'
 import { sanitize } from 'string-sanitizer'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
