+++ next/utils/blogpost.ts
@@ -1,4 +1,4 @@
-import { client } from './gql'
+import { client } from '../backend/utils/gql'
 
 export const fetchBlogPosts = async (tags: string[], limit: number, start: number) => {
   const { blogPosts } = await client.LatestPostsByTags({
