--- next/utils/regex.ts
@@ -1,1 +1,0 @@
-export const phoneRegex = /(^(\+\d+|0)?(?: ?\d{3}){3}$)/
